Source Stripe: Convert to airbyte-lib (#33940)
