+++ airbyte-integrations/connectors/source-stripe/main.py
@@ -3,43 +3,7 @@
 #
 
 
-import sys
-import traceback
-from datetime import datetime
-from typing import List
-
-from airbyte_cdk.entrypoint import AirbyteEntrypoint, launch
-from airbyte_cdk.models import AirbyteErrorTraceMessage, AirbyteMessage, AirbyteTraceMessage, TraceType, Type
-from source_stripe import SourceStripe
-
-
-def _get_source(args: List[str]):
-    catalog_path = AirbyteEntrypoint.extract_catalog(args)
-    config_path = AirbyteEntrypoint.extract_config(args)
-    try:
-        return SourceStripe(
-            SourceStripe.read_catalog(catalog_path) if catalog_path else None,
-            SourceStripe.read_config(config_path) if config_path else None,
-        )
-    except Exception as error:
-        print(
-            AirbyteMessage(
-                type=Type.TRACE,
-                trace=AirbyteTraceMessage(
-                    type=TraceType.ERROR,
-                    emitted_at=int(datetime.now().timestamp() * 1000),
-                    error=AirbyteErrorTraceMessage(
-                        message=f"Error starting the sync. This could be due to an invalid configuration or catalog. Please contact Support for assistance. Error: {error}",
-                        stack_trace=traceback.format_exc(),
-                    ),
-                ),
-            ).json()
-        )
-        return None
-
+from source_stripe.run import run
 
 if __name__ == "__main__":
-    _args = sys.argv[1:]
-    source = _get_source(_args)
-    if source:
-        launch(source, _args)
+    run()
