+++ docs/integrations/sources/stripe.md
@@ -216,6 +216,7 @@ Each record is marked with `is_deleted` flag when the appropriate event happens 
 
 | Version | Date       | Pull Request                                                 | Subject                                                                                                                                                                                                                       |
 |:--------|:-----------|:-------------------------------------------------------------|:------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
+| 5.1.2 | 2024-01-04 | [33414](https://github.com/airbytehq/airbyte/pull/33414) | Prepare for airbyte-lib |
 | 5.1.1   | 2024-01-04 | [33926](https://github.com/airbytehq/airbyte/pull/33926/)    | Update endpoint for `bank_accounts` stream                                                                                                                                                                                      |
 | 5.1.0   | 2023-12-11 | [32908](https://github.com/airbytehq/airbyte/pull/32908/)    | Read full refresh streams concurrently                                                                                                                                                                                        |
 | 5.0.2   | 2023-12-01 | [33038](https://github.com/airbytehq/airbyte/pull/33038)  | Add stream slice logging for SubStream                                                                                                                                                                                        |
